fix(nvim): get terraform fmt working properly
The default configuration only searches roots for .terraform or .git:

```
local util = require 'lspconfig.util'

return {
  default_config = {
    cmd = { 'terraform-lsp' },
    filetypes = { 'terraform', 'hcl' },
    root_dir = util.root_pattern('.terraform', '.git'),
  },
}
```

Because of this, it looks at the ENTIRE repo (:LspInfo will show the
root) if you are doing something like `vi ../../../../myfile.tf` instead
of reading a file in your current dir that has `.terraform` from when
you ran `init`.

Because of this, in large repos, formatting does not work if you are
editing a module when working in a state.

When editing a file, even using relative path, nvim sees the entire
filename as an absolute, so we just tell terraformls and terraform-lsp
that their `root` is the parent dir of the file we are, and `terraform
fmt` on save works once again.

The other change is to enable format on save and quit, which now works
in the newest version with the simple change to the lsp-format
configuration.

Additionally, add git blame and run linting/formatting

Signed-off-by: nomeelnoj <nomeelnoj@users.noreply.github.com>